requirements.txt
numpy
pandas
matplotlib>=3.1
pathlib==1.0.1
pyvisa==1.8